perf: make homepage static
To reduce initial loading time for users visiting through the homepage, the page is now statically built, with a revalidation every 24 hours.
Given the rate at which data is updated (not often), this should provide a decent experience.
Might need to revisit this whenever we start to allow users to update movies, as the updated-at slider will likely need some manual revalidation.