fix(auth): redirect in finally
Next's redirect apparently throws an error even when redirecting properly. As such, it needs to be called either outside of a try..catch or in finally. Here, we move the authentication redirect to finally, so we can still handle login errors but redirect either to the root or to the callback URL.